Integrate LLVM at llvm/llvm-project@3a82a1c3
Updates LLVM usage to match llvm/llvm-project@3a82a1c3. Further updates
the StableHLO submodule to openxla/stablehlo@2025586.

+++ test/Conversion/tosa-to-emitc.mlir
@@ -83,8 +83,8 @@ func.func @test_reciprocal(%arg0: tensor<13x21x3xf32>) -> tensor<13x21x3xf32> {
 }
 
 func.func @test_rescale(%arg0: tensor<13x21x3xui8>) -> tensor<13x21x3xi8> {
-  // CHECK: %0 = emitc.call_opaque "emitc::tosa::rescale"(%arg0) {args = [0 : index, 127 : i32, -1 : i32, array<i32: 1073741824>, array<i32: 30>, true, false, false], template_args = [tensor<13x21x3xi8>, 1 : i32]} : (tensor<13x21x3xui8>) -> tensor<13x21x3xi8>
-  %0 = "tosa.rescale"(%arg0) {double_round = false, input_zp = 127 : i32, multiplier = array<i32: 1073741824>, output_zp = -1 : i32, per_channel = false, scale32 = true, shift = array<i32: 30>} : (tensor<13x21x3xui8>) -> tensor<13x21x3xi8>
+  // CHECK: %0 = emitc.call_opaque "emitc::tosa::rescale"(%arg0) {args = [0 : index, 127 : i32, -1 : i32, array<i32: 1073741824>, array<i8: 30>, true, false, false], template_args = [tensor<13x21x3xi8>, 1 : i32]} : (tensor<13x21x3xui8>) -> tensor<13x21x3xi8>
+  %0 = "tosa.rescale"(%arg0) {double_round = false, input_zp = 127 : i32, multiplier = array<i32: 1073741824>, output_zp = -1 : i32, per_channel = false, scale32 = true, shift = array<i8: 30>} : (tensor<13x21x3xui8>) -> tensor<13x21x3xi8>
   return %0 : tensor<13x21x3xi8>
 }
